+++ src/handlers/handlers.js
@@ -1,4 +1,4 @@
-import {theList, inputField } from '../init/index.js'
+import { createListItem, sendUrl } from '../data.js'
 
 // control list through buttons 
 
@@ -7,53 +7,25 @@ const buttonControl = (event) => {
 
     event.preventDefault();
 
-const text = event.target.form.text.value;
+const text = event.target.form.text.value; // input field
 const action = event.target.value;
 
 if (action === "add") {
-
-// console.log('hey'); 
-
-// Div for each item
-const itemDiv = document.createElement('div');
-itemDiv.classList.add('item');
-theList.appendChild(itemDiv);
-
-//li for each item
-
-const liItem = document.createElement('li');
-liItem.innerText = inputField.value;
-liItem.classList.add('li-item');
-inputField.value = '';
-itemDiv.appendChild(liItem);
-
-//delete button 
-
-const deleteButton = document.createElement('button');
-deleteButton.innerHTML = '<i class = "fas fa-trash"></i>';
-// deleteButton.setAttribute('id', 'delete-button');
-deleteButton.classList.add('delete-button');
-itemDiv.appendChild(deleteButton);
-deleteButton.addEventListener('click', deleteItem);
+//   console.log(createListItem); 
+  createListItem(); 
 
 } 
-
-// send 
-
+// copy URLS 
 if (action === 'send') {
-
-    console.log('send');
-
+  
+  sendUrl();
 }
-
 // save
 
 if (action === 'save') {
 
     console.log('save');
-
 }
-
 
 //clean 
 
@@ -65,14 +37,8 @@ if (action === 'clean') {
 
 }
 
-
-
-
-
 const deleteItem = (event) => {
-    
     const item = event.target;
-
     // delete item
     if(item.classList[0] === 'delete-button') {
     const delItem = item.parentElement;
@@ -84,5 +50,4 @@ const deleteItem = (event) => {
     
 }
 
-
-export {buttonControl, deleteItem}
+export { buttonControl, deleteItem }
