+++ src/listeners/listener.js
@@ -1,15 +1,21 @@
 
 import { addButton, sendButton, saveButton, cleanButton } from '../init/index.js'
 import { buttonControl } from '../handlers/handlers.js'
-// import { deleteItem } from '../handlers/handlers.js'
+// import { saveLocalItem } from '../data.js'
+
+
+// Button control listeners
 
 addButton.addEventListener('click', buttonControl);
-
-
-// const buttonAdd = document.createElement('button');
 
 sendButton.addEventListener('click', buttonControl);
 
 saveButton.addEventListener('click', buttonControl);
 
 cleanButton.addEventListener('click', buttonControl);
+
+// Local storage 
+
+// addButton.addEventListener('DOMContentLoaded', saveLocalItem);
+
+//
